Izgenerirane skripte, potrebno jos upute u njima napisat za argumente

+++ lab3/src/main/scala/hr/vinko/nos/lab3/FieldUtil.scala
@@ -3,7 +3,7 @@ package hr.vinko.nos.lab3
 object FieldUtil {
 
   val MAX_COL = 60
-  
+
   val HEADER = "---BEGIN OS2 CRYPTO DATA---"
   val FOOTER = "---END OS2 CRYPTO DATA---"
 
@@ -14,9 +14,9 @@ object FieldUtil {
   def createField(title: String, text: String): String = {
     createField(title, split(text, MAX_COL))
   }
-  
+
   def createField(title: String, text: List[String]): String = {
-     val sb = new StringBuilder
+    val sb = new StringBuilder
 
     sb.append(title + "\n")
 
@@ -34,11 +34,15 @@ object FieldUtil {
     if (text.size <= n) text :: Nil
     else (text take n) :: split(text drop n, n)
   }
-  
-  def getField(title: String, lines: List[String]) = {
+
+  def getField(title: String, lines: List[String]): String = {
     val index = lines.indexOf(title + ":")
-    val tmp = lines drop index+1
-    (tmp take tmp.indexOf("")).toList map { _.trim } mkString("")
+    if (index >= 0) {
+      val tmp = lines drop index + 1
+      (tmp take tmp.indexOf("")).toList map { _.trim } mkString ("")
+    } else {
+      "0"
+    }
   }
 
   def createDescription(text: String) = {
@@ -52,19 +56,19 @@ object FieldUtil {
   def createMethod(text: String) = {
     createField("Method:", text)
   }
-  
+
   def createMethod(text: List[String]) = {
     createField("Method:", text)
   }
-  
+
   def createKeyLength(text: String) = {
     createField("Key length:", text)
   }
-  
+
   def createKeyLength(text: List[String]) = {
     createField("Key length:", text)
   }
-  
+
   def createSecretKey(text: String) = {
     createField("Secret key:", text)
   }
